+++ js/app/controllers/sidebar/playlistdetailscontroller.js
@@ -43,7 +43,7 @@ angular.module('Music').controller('PlaylistDetailsController', [
 
 		$scope.$watchCollection('playlist.tracks', function() {
 			$scope.totalLength = _.reduce($scope.playlist.tracks, function(sum, item) {
-				return sum + (item.track ? item.track.length : 0); // be prepared for invalid playist entries
+				return sum + (item.track?.length ?? 0); // be prepared for invalid playist entries
 			}, 0);
 		});